pyproject.toml
[tool.poetry]
name = "config-synth-flow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
pydantic = "^2.10.6"
tqdm = "^4.67.1"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
ipython = "^9.0.2"
litellm = "^1.63.11"
datasets = "^3.4.0"
transformers = "^4.49.0"
langchain-text-splitters = "^0.3.6"
opencc = "^1.1.9"
qdrant-client = "^1.13.3"
hanzidentifier = "^1.3.0"

[tool.ruff]
fix = true
include = ["config_synth_flow/**"]
line-length = 100

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
docstring-code-line-length = 100
exclude = ["*pyc"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.0"
pytest-asyncio = "^0.25.3"
pytest-timeout = "^2.3.1"
fire = "^0.7.0"


[tool.poetry.group.dedup.dependencies]
rensa = "^0.2.0"
semhash = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run-seq = "config_synth_flow.scripts.run_seq:main"